Moved JVM specific Blocker constructors to JVM only source directory
