+++ core/shared/src/main/scala/cats/effect/Blocker.scala
@@ -18,8 +18,8 @@ package cats
 package effect
 
 import scala.concurrent.ExecutionContext
-import java.util.concurrent.{ExecutorService, Executors, ThreadFactory}
-import cats.data.NonEmptyList
+
+import cats.effect.internals.BlockerPlatform
 
 /**
  * An execution context that is safe to use for blocking operations.
@@ -46,50 +46,7 @@ final class Blocker private (val blockingContext: ExecutionContext) {
     cs.evalOn(this)(fa)
 }
 
-object Blocker {
-
-  /**
-   * Creates a blocker that is backed by a cached thread pool.
-   */
-  def apply[F[_]](implicit F: Sync[F]): Resource[F, Blocker] =
-    fromExecutorService(F.delay(Executors.newCachedThreadPool(new ThreadFactory {
-      def newThread(r: Runnable) = {
-        val t = new Thread(r, "cats-effect-blocker")
-        t.setDaemon(true)
-        t
-      }
-    })))
-
-  /**
-   * Creates a blocker backed by the `ExecutorService` returned by the
-   * supplied task. The executor service is shut down upon finalization
-   * of the returned resource.
-   * 
-   * If there are pending tasks in the thread pool at time the returned
-   * `Blocker` is finalized, the finalizer fails with a `Blocker.OutstandingTasksAtShutdown` 
-   * exception.
-   */
-  def fromExecutorService[F[_]](makeExecutorService: F[ExecutorService])(implicit F: Sync[F]): Resource[F, Blocker] =
-    Resource.make(makeExecutorService) { ec =>
-        val tasks = F.delay {
-          val tasks = ec.shutdownNow()
-          val b = List.newBuilder[Runnable]
-          val itr = tasks.iterator
-          while (itr.hasNext)
-            b += itr.next
-          NonEmptyList.fromList(b.result)
-        }
-        F.flatMap(tasks) {
-          case Some(t) => F.raiseError(new OutstandingTasksAtShutdown(t))
-          case None => F.unit
-        }
-    }.map(es => liftExecutorService(es))
-
-  /**
-   * Creates a blocker that delegates to the supplied executor service.
-   */
-  def liftExecutorService(es: ExecutorService): Blocker =
-    liftExecutionContext(ExecutionContext.fromExecutorService(es))
+object Blocker extends BlockerPlatform {
 
   /**
    * Creates a blocker that delegates to the supplied execution context.
@@ -98,7 +55,4 @@ object Blocker {
    * `scala.concurrent.ExecutionContext.Implicits.global'.
    */
   def liftExecutionContext(ec: ExecutionContext): Blocker = new Blocker(ec)
-
-  /** Thrown if there are tasks queued in the thread pool at the time a `Blocker` is finalized. */
-  final class OutstandingTasksAtShutdown(val tasks: NonEmptyList[Runnable]) extends IllegalStateException("There were outstanding tasks at time of shutdown of the thread pool")
 }
